requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.6
django-import-export==3.0.2
et-xmlfile==1.1.0
executing==1.2.0
forex-python==1.8
gunicorn==20.1.0
idna==3.4
ipykernel==6.21.1
ipython==8.9.0
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
MarkupPy==1.14
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.3.0
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
whitenoise==6.3.0
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
